+++ lib/features/app/presentation/widgets/appbar/color_animated_icon.dart
